+++ jax/_src/test_util.py
@@ -1230,18 +1230,21 @@ def parameterized_filterable(*,
   Args:
     kwargs: Each entry is a set of kwargs to be passed to the test function.
     testcase_name: Optionally, a function to construct the testcase_name from
-      one kwargs dict. If not given then the kwarg must contain `testcase_name`.
+      one kwargs dict. If not given then kwarg may contain `testcase_name` and
+      if not, the test case name is constructed as `str(kwarg)`.
     one_containing: If given, then leave the test name unchanged, and use
       only one `kwargs` whose `testcase_name` includes `one_containing`.
   """
   # Ensure that all kwargs contain a testcase_name
   kwargs_with_testcase_name: Sequence[dict[str, Any]]
   if testcase_name is not None:
-    kwargs_with_testcase_name = [dict(testcase_name=testcase_name(kw), **kw)
+    kwargs_with_testcase_name = [dict(testcase_name=str(testcase_name(kw)), **kw)
                                  for kw in kwargs]
   else:
     for kw in kwargs:
-      assert "testcase_name" in kw
+      if "testcase_name" not in kw:
+        kw["testcase_name"] = "_".join(f"{k}={str(kw[k])}"
+                                       for k in sorted(kw.keys()))
     kwargs_with_testcase_name = kwargs
   if one_containing is not None:
     filtered = tuple(kw for kw in kwargs_with_testcase_name